Tkinter/MauroCespedes_TkinterAPP/requirements.txt
altgraph==0.17.2
cycler==0.11.0
fonttools==4.33.3
future==0.18.2
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pefile==2021.9.3
Pillow==9.1.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tk==0.1.0
